+++ .gitignore
@@ -1,4 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+secrets.json
 
 # dependencies
 /node_modules
