+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "axios": "^1.3.0",
+    "googleapis": "^110.0.0",
     "next": "13.1.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
