--- pages/activities/index.js
@@ -1,79 +1,0 @@
-import React from "react";
-import Hero from "../../comps/Hero";
-import ImgSrc from "../../public/recroom-hero.jpg";
-import Head from "next/head";
-
-const ActivityPage = () => {
-  return (
-    <>
-      <Head>
-        <title>Activity Page</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/icon.png" />
-        <link
-          href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@200;400;500&family=Russo+One&display=swap"
-          rel="stylesheet"
-        />
-      </Head>
-
-      <Hero img={ImgSrc} />
-      <div class="flex justify-center w-full border-2 border-green-500">
-        <div class="flex w-3/4 p-5 border-2 border-red-500">
-          {/* 2 columns of activity data */}
-          <div class="leftDiv flex flex-grow flex-col p-4 px-24 border-2 border-purple-300">
-            <h1 class="text-5xl font-montserrat font-bold">Rec Room - Amazing Brentwood</h1>
-            <p class="font-montserrat text-2xl leading-8">
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-              eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-              enim ad minim veniam, quis nostrud exercitation ullamco laboris
-              nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in
-              reprehenderit in voluptate velit esse cillum dolore eu fugiat
-              nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-              sunt in culpa qui officia deserunt mollit anim id est laborum.
-            </p>
-            <h2 class="font-montserrat font-bold text-2xl leading-7">Subheader</h2>
-            <p class="font-montserrat text-2xl leading-7">
-              "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-              eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-              enim ad minim veniam, quis nostrud exercitation ullamco laboris
-              nisi ut aliquip ex ea commodo consequat.
-            </p>
-          </div>
-
-          <div class="rightDiv flex flex-col">
-            <div class="flex flex-col p-10 bg-gray-300 mb-5">
-              <div class="my-2">
-                <h3 class="font-bold">Activity Type:</h3>
-                <p>Entertainment</p>
-              </div>
-
-              <div class="my-2">
-                <h3 class="font-bold">Activity Type:</h3>
-                <p>Entertainment</p>
-              </div>
-
-              <div class="my-2">
-                <h3 class="font-bold">Activity Type:</h3>
-                <p>Entertainment</p>
-              </div>
-            </div>
-
-            <iframe
-              src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d83327.51117460211!2d-123.123904!3d49.25773545!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x548673f143a94fb3%3A0xbb9196ea9b81f38b!2sVancouver%2C%20BC!5e0!3m2!1sen!2sca!4v1675211148040!5m2!1sen!2sca"
-              width="300"
-              height="300"
-              style={{ border: 0 }}
-              allowfullscreen=""
-              loading="lazy"
-              referrerpolicy="no-referrer-when-downgrade"
-            ></iframe>
-
-            <button class="w-full bg-gray-400 py-3 mt-5">Add to list</button>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default ActivityPage;
